backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9